update! change crossgarbage table name to detailtransaction

+++ app/src/main/java/com/bangkit/ecoease/data/model/Order.kt
@@ -2,10 +2,10 @@ package com.bangkit.ecoease.data.model
 
 data class Order(
     val garbageList: List<GarbageAdded?>,
-    val total: Int
+    val total: Long
 )
 data class GarbageAdded(
     val garbage: com.bangkit.ecoease.data.room.model.Garbage,
     val amount: Int,
-    val totalPrice: Int
+    val totalPrice: Long
 )
